+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#27]: Add `simpleValue` property to disable the default behavior of returning an array of values rather than an array of option objects
 * [#28]: Add `availableLabel` and `selectedLabel` to change the hidden control label texts
+* [#30]: Add `moveKeyCodes` property to set the key codes that trigger a move for the select options
 
 ## [v1.2.0](https://github.com/jakezatecky/react-dual-listbox/compare/v1.1.0...v1.2.0) (2017-08-12)